[HUDI-1785] Move OperationConverter to hudi-client-common for code reuse
